+++ libs/attachment-canvas/src/components/VisualizerCanvasRenderer/VisualizerCanvasRenderer.module.scss
@@ -1,11 +1,11 @@
 .loadingOverlay {
-  background-color: var(--bg-layer-1, #e0e6f0);
+  background-color: var(--vs-loading-bg, var(--bg-layer-1, #e0e6f0));
 }
 
 .statusLabel {
-  color: var(--text-primary, #161b2d);
+  color: var(--vs-status-text, var(--text-primary, #161b2d));
 }
 
 .errorIcon {
-  color: var(--text-error, #ae2f2f);
+  color: var(--vs-error-icon, var(--text-error, #ae2f2f));
 }
