+++ libs/attachment-canvas/tsconfig.lib.json
@@ -38,9 +38,6 @@
     },
     {
       "path": "../sidebar/tsconfig.lib.json"
-    },
-    {
-      "path": "../visualizer-connector/tsconfig.lib.json"
     }
   ]
 }
