+++ libs/chat-shared/src/index.ts
@@ -8,12 +8,11 @@ export * from './models/deployment';
 export * from './models/deployment-features';
 export * from './models/import-export';
 export * from './models/tool-menu-item';
+export * from './models/custom-visualizer';
 export * from './types/attachment';
 export * from './types/mime-type';
 export * from './types/code-editor';
 export * from './types/overlay/overlay-protocol';
-export * from './types/visualizer-connector';
-export * from './constants/visualizer-connector';
 export * from './utils/string-utils';
 export * from './utils/merge-class';
 export * from './utils/build-css-vars';
