+++ libs/chat-shared/src/models/custom-visualizer.ts
@@ -60,59 +60,3 @@ export interface CustomVisualizerData {
   /** Opaque attachment payload consumed by the third-party visualizer. */
   [key: string]: unknown;
 }
-
-/** Full `SEND_VISUALIZE_DATA` request payload. */
-export interface AttachmentData {
-  /** MIME type of the attachment being visualized. */
-  mimeType: string;
-  /** Layout hints plus the opaque attachment payload. */
-  visualizerData: CustomVisualizerData;
-}
-
-/** A single attachment entry inside a `SEND_GROUPED_VISUALIZE_DATA` payload. */
-export interface AttachmentItem {
-  /** URL of the attachment resource. */
-  url: string;
-  /** MIME type of the attachment. */
-  mimeType: string;
-  /** Layout hints plus opaque payload for this specific attachment. */
-  visualizerData: CustomVisualizerData;
-}
-
-/** Full `SEND_GROUPED_VISUALIZE_DATA` request payload for application-level visualizers. */
-export interface GroupedAttachmentsData {
-  /** Ordered list of attachments to render together. */
-  attachments: AttachmentItem[];
-  /** Shared layout hints applied to the whole group. */
-  layout: CustomVisualizerDataLayout;
-}
-
-/** Options accepted by the host-side `VisualizerConnector` constructor. */
-export interface VisualizerConnectorOptions {
-  /** URL of the visualizer application to load in the iframe. */
-  domain: string;
-  /**
-   * Wire-format namespace prefix used on every outbound and inbound message
-   * (`${visualizerName}/…`). Must equal the iframe-side application's
-   * `appName` or no messages will be recognized on either side.
-   */
-  visualizerName: string;
-  /** Milliseconds to wait for a `send()` request's `/RESPONSE` before rejecting. Defaults to `10000`. Does NOT bound `ready()`. */
-  requestTimeout?: number;
-  /** CSS styles applied to the built-in loader overlay element while the iframe is not yet ready. */
-  loaderStyles?: Record<string, string>;
-  /** CSS class(es) applied to the built-in loader overlay element. */
-  loaderClass?: string;
-  /** HTML string set as the `innerHTML` of the built-in loader overlay element. */
-  loaderInnerHTML?: string;
-}
-
-/** Wire envelope for every message exchanged between host and visualizer iframe. */
-export interface VisualizerConnectorRequest {
-  /** Full wire type string, `${visualizerName}/${eventOrRequestName}` (or `.../RESPONSE`). */
-  type: string;
-  /** Present on request/response envelopes; absent on unsolicited event envelopes. */
-  requestId?: string;
-  /** Envelope-specific payload. */
-  payload?: unknown;
-}
