--- libs/chat-visualizer-connector/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@epam/ai-dial-chat-visualizer-connector",
-  "version": "0.0.1",
-  "description": "Iframe-side postMessage connector for third-party visualizer applications to receive attachment data from DIAL Chat",
-  "license": "Apache-2.0",
-  "private": true,
-  "type": "module",
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "@epam/source": "./src/index.ts",
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "default": "./dist/index.js"
-    }
-  },
-  "files": ["dist", "!**/*.tsbuildinfo"],
-  "peerDependencies": {
-    "@epam/ai-dial-chat-shared": "*"
-  },
-  "nx": {
-    "name": "chat-visualizer-connector",
-    "tags": ["publishable"],
-    "targets": {
-      "publish": {
-        "command": "node tools/publish-lib.mjs {projectName} --version={args.ver} --dry={args.dry} --tag={args.tag} --development={args.development}"
-      }
-    }
-  },
-  "dependencies": {}
-}
