--- libs/chat-visualizer-connector/src/index.ts
@@ -1,16 +1,0 @@
-export {
-  ChatVisualizerConnector,
-  type ChatVisualizerCallbacks,
-  type PostMessageRequestParams,
-} from './lib/ChatVisualizerConnector';
-
-export type {
-  AttachmentData,
-  AttachmentItem,
-  CustomVisualizerData,
-  GroupedAttachmentsData,
-} from '@epam/ai-dial-chat-shared';
-export {
-  VisualizerConnectorEvents,
-  VisualizerConnectorRequests,
-} from '@epam/ai-dial-chat-shared';
