--- libs/visualizer-connector/src/index.ts
@@ -1,14 +1,0 @@
-export { VisualizerConnector } from './lib/VisualizerConnector';
-
-export type {
-  AttachmentData,
-  CustomVisualizer,
-  CustomVisualizerData,
-  CustomVisualizerDataLayout,
-  VisualizerConnectorOptions,
-  VisualizerConnectorRequest,
-} from '@epam/ai-dial-chat-shared';
-export {
-  VisualizerConnectorEvents,
-  VisualizerConnectorRequests,
-} from '@epam/ai-dial-chat-shared';
