+++ package.json
@@ -100,8 +100,10 @@
   },
   "dependencies": {
     "@epam/ai-dial-react-pdf-highlighter": "0.2.0-dev.22",
+    "@epam/ai-dial-shared": "0.48.0",
     "@epam/ai-dial-typescript-sdk": "0.1.0-dev.31",
     "@epam/ai-dial-ui-kit": "^0.13.0-dev.17",
+    "@epam/ai-dial-visualizer-connector": "0.48.0",
     "@epam/pdf-highlighter-kit": "^0.0.18",
     "@nestjs/cache-manager": "^3.1.2",
     "@nestjs/common": "^11.0.0",
