+++ tsconfig.json
@@ -65,12 +65,6 @@
     },
     {
       "path": "./libs/publish-panel"
-    },
-    {
-      "path": "./libs/visualizer-connector"
-    },
-    {
-      "path": "./libs/chat-visualizer-connector"
     }
   ]
 }
